Switch to Typescript

+++ src/lib/courseChannels.ts
@@ -1,79 +1,83 @@
-import { logger } from "../logger.js";
 import _ from "lodash";
-import { Course } from "../models/course.js";
-import { CourseRolesSetting } from "../models/configuration_models/courseRolesSetting.js";
-import { messageEmbed } from "./messageEmbed.js";
 import colorConvert from "color-convert";
+import { logger } from "../logger";
+import { Course } from "../models/course";
+import { CourseRolesSetting } from "../models/configuration_models/courseRolesSetting";
+import { messageEmbed } from "./messageEmbed";
+import type { ColorResolvable, CommandInteraction, Guild, GuildMemberResolvable } from "discord.js";
+import { HexColorString } from "discord.js";
 
 //import {inspect} from "util";
 
 /**
- * @param {Client} client The Discord bots client.
- * @param {User} user The Discord user in question.
+ * @param {Guild} guild The Discord guild.
+ * @param {GuildMemberResolvable} user The Discord user in question.
  * @param {Course} course The database entry for the course.
  */
-export async function addUserToCourseChannel(guild, user, course) {
+export async function addUserToCourseChannel(guild: Guild, user: GuildMemberResolvable, course: Course) {
     const targetUser = guild.members.resolve(user);
-    await targetUser.roles.add(course.roleId);
+    await targetUser?.roles.add(course.roleId);
 }
 
 /**
- * @param {Client} client The Discord bots client.
+ * @param {Guild} guild The Discord guild
  * @param {User} user The Discord user in question.
  * @param {Course} course The database entry for the course.
  */
-export async function removeUserFromCourseChannel(guild, user, course) {
+export async function removeUserFromCourseChannel(guild: Guild, user: GuildMemberResolvable, course: Course) {
     const targetUser = guild.members.resolve(user);
-    await targetUser.roles.remove(course.roleId);
+    await targetUser?.roles.remove(course.roleId);
 }
 
 /**
- * @param {Client} client The Discord bots client.
  * @param {string} name The name of the course and the course role.
- * @param {Guild} guild The guild in which the channel is being created for.
+ * @param {CommandInteraction} interaction
  */
-export async function createCourseChannel(name, interaction) {
+export async function createCourseChannel(name: string, interaction: CommandInteraction) {
     const guild = interaction.guild;
 
-    const courseRoleSettings = await CourseRolesSetting.findOne({ where: { guildId: guild.id } });
+    if (!guild) {
+        throw new Error('Guild not found.');
+    }
+
+    const courseRoleSettings = await CourseRolesSetting.findOne({ where: { guildId: guild?.id } });
     if(courseRoleSettings === null) {
         interaction.reply({ embeds: [messageEmbed("Course role settings are not configured.", "RED")]});
     }
 
-    const joinMessageChannel = await guild.channels.fetch(courseRoleSettings.roleSelectionChannelId);
+    const joinMessageChannel = await guild?.channels.fetch(courseRoleSettings?.roleSelectionChannelId!);
     
-    // FIXME: This is throwing an error! Why? I really don't know... isTextBased is in fact a documented function and yet it doesn't want to work.
-    /*if(joinMessageChannel.isTextBased() === false) {
+    if (!joinMessageChannel?.isText()) {
         throw new Error("Could not send message to join message channel as it is not text based.");
-    }*/
+    }
 
     name = _.upperCase(name);
     name = name.replace(/ /g, "");
 
-    const role = await guild.roles.create({ name });
-    role.setHoist(true);
-    role.setMentionable(true);
+    const role = await guild?.roles.create({ name });
+    role?.setHoist(true);
+    role?.setMentionable(true);
 
 
-    const courseChannel = await guild.channels.create(name);
-    await courseChannel.setParent(courseRoleSettings.courseChatCategoryId, { lockPermissions: false });
-    await courseChannel.permissionOverwrites.edit(role, {
+    const courseChannel = await guild?.channels.create(name);
+    await courseChannel?.setParent(courseRoleSettings?.courseChatCategoryId!, { lockPermissions: false });
+    await courseChannel?.permissionOverwrites.edit(role!, {
         VIEW_CHANNEL: true
     });
 
-    await courseChannel.permissionOverwrites.edit(guild.roles.everyone, {
+    await courseChannel?.permissionOverwrites.edit(guild?.roles.everyone!, {
         VIEW_CHANNEL: false
     });
 
-    const joinMessage = await joinMessageChannel.send(`**${name}** <@&${role.id}>`);
+    const joinMessage = await joinMessageChannel?.send(`**${name}** <@&${role?.id}>`);
     // **CS487** @CS487
     await joinMessage.react('👍');
 
     // Create the database entry
     await Course.create({
         name: name,
-        channelId: courseChannel.id,
-        roleId: role.id,
+        channelId: courseChannel?.id!,
+        roleId: role?.id!,
         messageId: joinMessage.id
     });
 
@@ -85,9 +89,10 @@ export async function createCourseChannel(name, interaction) {
  * Adds an existing course channel to the database.
  * @param {String} channelId 
  * @param {String} joinMessageId 
- * @param {String} roleId 
+ * @param {String} roleId
+ * @param {String} name
  */
-export async function linkExistingCourseChannel(channelId, joinMessageId, roleId, name) {
+export async function linkExistingCourseChannel(channelId: string, joinMessageId: string, roleId: string, name: string) {
     await Course.create({
         name: name,
         channelId: channelId,
@@ -99,8 +104,9 @@ export async function linkExistingCourseChannel(channelId, joinMessageId, roleId
 /**
  * Deletes the course channel and related roles.
  * @param {String} roleId 
+ * @param {CommandInteraction} interaction
  */
-export async function deleteCourseChannel(roleId, interaction) {
+export async function deleteCourseChannel(roleId: string, interaction: CommandInteraction) {
     const guild = interaction.guild;
     const course = await Course.findOne({
         where: { roleId: roleId }
@@ -112,7 +118,7 @@ export async function deleteCourseChannel(roleId, interaction) {
     }
 
     const courseRoleSetting = await CourseRolesSetting.findOne({
-        where: { guildId: guild.id }
+        where: { guildId: guild?.id }
     });
 
     if(!courseRoleSetting) {
@@ -121,23 +127,27 @@ export async function deleteCourseChannel(roleId, interaction) {
     }
 
     const joinChannelId = courseRoleSetting.roleSelectionChannelId;
-    const joinChannel = await guild.channels.fetch(joinChannelId);
+    const joinChannel = await guild?.channels.fetch(joinChannelId);
+
+    if (!joinChannel?.isText()) {
+        throw new Error("Could not send message to join channel as channel is not text-based.");
+    }
 
     const messageId = course.messageId;
     const joinMessage = await joinChannel.messages.fetch(messageId);
     await joinMessage.delete();
 
     const channelId = course.channelId;
-    const channel = await guild.channels.fetch(channelId);
-    await channel.delete();
+    const channel = await guild?.channels.fetch(channelId);
+    await channel?.delete();
 
-    const role = await guild.roles.fetch(roleId);
-    await role.delete();
+    const role = await guild?.roles.fetch(roleId);
+    await role?.delete();
 
     await course.destroy();
 }
 
-export async function unlinkExistingCourseChannel(roleId, interaction) {
+export async function unlinkExistingCourseChannel(roleId: string, interaction: CommandInteraction) {
     const course = await Course.findOne({
         where: { roleId: roleId }
     });
@@ -153,7 +163,7 @@ export async function unlinkExistingCourseChannel(roleId, interaction) {
 /**
  * Updates the colors of the course roles in the server
  */
-export function updateCourseColors(guild) {
+export function updateCourseColors(guild: Guild) {
     return new Promise((resolve, reject) => {
         Course.findAll().then(courses => {
             const courseNames = courses.map(course => course.name);
@@ -168,7 +178,7 @@ export function updateCourseColors(guild) {
                 if (!color) {
                     reject(`No color found for course ${course.name}`);
                 }
-                return role.setColor(color);
+                return role?.setColor(color);
             });
         
             Promise.all(roleUpdates)
@@ -184,11 +194,7 @@ export function updateCourseColors(guild) {
     });
 }
 
-/**
- * Extracts the first number from a string.
- * @param {String} courseName
- */
-function getCourseColors(courses) {
+function getCourseColors(courses: string[]) {
     const sortedClassNames = courses.sort();
 
     // define the start and end colors of the gradient
@@ -198,16 +204,13 @@ function getCourseColors(courses) {
     const lightness = 50;
 
     // compute the color gradient
-    const numClasses = sortedClassNames.length;
-    const colors = [];
-    for (let i = 0; i < numClasses; i++) {
-        // compute the hue for this class
-        const hue = startHue + (endHue - startHue) * i / (numClasses - 1);
-        const hsl = [hue, saturation, lightness];
-        colors.push(hsl);
-    }
+    const colorMap: Record<string, ColorResolvable> = {};
 
-    // map the sorted class names to their corresponding colors
-    const colorMap = Object.fromEntries(sortedClassNames.map((name, i) => [name, colorConvert.hsl.hex(colors[i])]));
+    sortedClassNames.forEach((className, index, coll) => {
+        const hue = startHue + (endHue - startHue) * index / (coll.length - 1);
+        const hsl: [number, number, number] = [hue, saturation, lightness];
+        colorMap[className] = colorConvert.hsl.hex(hsl) as HexColorString;
+    })
+
     return colorMap;
 }
